Check read buffer is large enough for all the nvm3 object data.
(nvm3 read throws an error if read buffer is not large enough for ALL the obj
data, so we have to make sure the buffer is big enough => partial read is
NOT supported like original ot nv application.  If this is still required
we will need to buffer nvm3 read data before write to output buffer.